book_search/requirements.txt
Django>=5.0
djangorestframework
python-decouple
psycopg2-binary
requests
beautifulsoup4
django-cors-headers
gunicorn